_BaseEnumField: Make run_validators method a no-op
See the comment in this commit-- I can't see value in allowing custom
validators on EnumFields and the implementation in the superclass causes
warnings in RichEnum.__eq__.

Arguably those warnings aren't useful (warning against []/falsy compare).
In that case, we can revert this when they're silenced.

Alternatively, if we need the warnings and need this functionality, we'd have
re-implement the method in the superclass without said check, or live with
warnings every time a form containing an EnumField is validated, which sucks.